Split patch adapter tests by catalog
