+++ Aikido.Zen.Test/PatcherTests.cs
@@ -1,16 +1,11 @@
 using System.Data.Common;
-using System.Diagnostics;
 using System.IO;
-using System.Net;
 using System.Net.Http;
 using System.Reflection;
 using System.Threading;
 using Aikido.Zen.Core;
-using Aikido.Zen.Core.Api;
 using Aikido.Zen.Core.Sinks;
-using Aikido.Zen.Tests.Mocks;
 using HarmonyLib;
-using Moq;
 
 namespace Aikido.Zen.Test
 {
@@ -19,7 +14,6 @@ namespace Aikido.Zen.Test
     {
         private const string HarmonyId = "aikido.zen";
         private Context _context = null!;
-        private Agent _agent = null!;
 
         [SetUp]
         public void SetUp()
@@ -29,42 +23,12 @@ namespace Aikido.Zen.Test
             {
                 ParsedUserInput = new Dictionary<string, string>()
             };
-
-            Environment.SetEnvironmentVariable("AIKIDO_TOKEN", "test-token");
-            Environment.SetEnvironmentVariable("AIKIDO_BLOCK", "true");
-            Environment.SetEnvironmentVariable("AIKIDO_URL", "http://localhost:3000");
-            Environment.SetEnvironmentVariable("AIKIDO_REALTIME_URL", "http://localhost:3001");
-
-            var reportingApiMock = new Mock<IReportingAPIClient>();
-            reportingApiMock
-                .Setup(r => r.ReportAsync(It.IsAny<string>(), It.IsAny<object>()))
-                .ReturnsAsync(new ReportingAPIResponse { Success = true });
-            reportingApiMock
-                .Setup(r => r.GetFirewallLists(It.IsAny<string>()))
-                .ReturnsAsync(new FirewallListsAPIResponse { Success = true });
-
-            var runtimeApiMock = new Mock<IRuntimeAPIClient>();
-            runtimeApiMock
-                .Setup(r => r.GetConfig(It.IsAny<string>()))
-                .ReturnsAsync(new ReportingAPIResponse { Success = true });
-            runtimeApiMock
-                .Setup(r => r.GetConfigLastUpdated(It.IsAny<string>()))
-                .ReturnsAsync(new ConfigLastUpdatedAPIResponse { Success = true });
-
-            _agent = Agent.NewInstance(ZenApiMock.CreateMock(reportingApiMock.Object, runtimeApiMock.Object).Object);
-            _agent.ClearContext();
         }
 
         [TearDown]
         public void TearDown()
         {
             Patcher.Unpatch();
-            _agent?.Dispose();
-
-            Environment.SetEnvironmentVariable("AIKIDO_TOKEN", null);
-            Environment.SetEnvironmentVariable("AIKIDO_BLOCK", null);
-            Environment.SetEnvironmentVariable("AIKIDO_URL", null);
-            Environment.SetEnvironmentVariable("AIKIDO_REALTIME_URL", null);
         }
 
         [Test]
@@ -135,85 +99,6 @@ namespace Aikido.Zen.Test
 
             Assert.That(OverloadFallbackTarget.Execute("value"), Is.EqualTo("one-argument"));
             Assert.That(OverloadFallbackTarget.Execute("value", 1), Is.EqualTo("patched-overload"));
-        }
-
-        [Test]
-        public void PatchCatalog_ForwardsIoArgumentsToSink()
-        {
-            Patcher.PatchSinks(() => _context);
-
-            Assert.That(IOPatches.OnePath(
-                "safe.txt",
-                GetMethod(typeof(File), nameof(File.ReadAllText), typeof(string))), Is.True);
-            Assert.That(IOPatches.TwoFilePaths(
-                "source.txt",
-                "destination.txt",
-                GetMethod(typeof(File), nameof(File.Copy), typeof(string), typeof(string), typeof(bool))), Is.True);
-            Assert.That(IOPatches.PathWithBasePath(
-                "child.txt",
-                Path.GetTempPath(),
-                GetMethod(typeof(Path), nameof(Path.GetFullPath), typeof(string), typeof(string))), Is.True);
-        }
-
-        [Test]
-        public void PatchCatalog_ForwardsOutboundArgumentsToSink()
-        {
-            Patcher.PatchSinks(() => _context);
-            var method = GetMethod(
-                typeof(HttpClient),
-                nameof(HttpClient.SendAsync),
-                typeof(HttpRequestMessage),
-                typeof(CancellationToken));
-
-            using (var httpClient = new HttpClient())
-            using (var request = new HttpRequestMessage(HttpMethod.Get, "https://safe.example/path"))
-            {
-                Assert.That(OutboundRequestPatches.HttpClientRequest(request, httpClient, method), Is.True);
-            }
-
-            using (var httpClient = new HttpClient { BaseAddress = new Uri("https://base.example") })
-            using (var request = new HttpRequestMessage(HttpMethod.Get, "/relative"))
-            {
-                Assert.That(OutboundRequestPatches.HttpClientRequest(request, httpClient, method), Is.True);
-            }
-
-            using (var httpClient = new HttpClient { BaseAddress = new Uri("https://base-only.example") })
-            {
-                Assert.That(OutboundRequestPatches.HttpClientRequest(null!, httpClient, method), Is.True);
-            }
-
-#pragma warning disable SYSLIB0014
-            var webRequest = WebRequest.Create("https://safe.example/path");
-#pragma warning restore SYSLIB0014
-            Assert.That(OutboundRequestPatches.WebRequest(webRequest, GetMethod(typeof(WebRequest), nameof(WebRequest.GetResponse))), Is.True);
-        }
-
-        [Test]
-        public void PatchCatalog_ForwardsSqlProcessAndLlmArgumentsToSinks()
-        {
-            Patcher.PatchSinks(() => _context);
-
-            var dbCommand = new Mock<DbCommand>();
-            dbCommand.SetupGet(command => command.CommandText).Returns("SELECT 1");
-            var dbMethod = GetMethod(typeof(DbCommand), nameof(DbCommand.ExecuteScalar));
-
-            Assert.That(SqlClientPatches.DbCommand(dbCommand.Object, dbMethod), Is.True);
-            Assert.That(SqlClientPatches.NPocoCommand(dbCommand.Object, dbMethod), Is.True);
-            Assert.That(SqlClientPatches.ExecuteSqlRaw(
-                "SELECT 1",
-                GetMethod(typeof(TestSqlMethods), nameof(TestSqlMethods.ExecuteSqlRaw), typeof(object), typeof(string), typeof(IEnumerable<object>))), Is.True);
-
-            using (var process = new Process { StartInfo = new ProcessStartInfo("echo", "safe") })
-            {
-                Assert.That(ProcessExecutionPatches.ProcessStart(
-                    process,
-                    GetMethod(typeof(Process), nameof(Process.Start))), Is.True);
-            }
-
-            Assert.DoesNotThrow(() => LLMPatches.LLMCallCompleted(
-                null!,
-                null!,
-                GetMethod(typeof(object), nameof(ToString))));
         }
 
         private static MethodInfo GetMethod(Type type, string methodName, params Type[] parameterTypes)
@@ -348,12 +233,5 @@ namespace Aikido.Zen.Test
             }
         }
 
-        private static class TestSqlMethods
-        {
-            public static int ExecuteSqlRaw(object databaseFacade, string sql, IEnumerable<object> parameters)
-            {
-                return 0;
-            }
-        }
     }
 }
